fixed bugs

+++ ex01/include/Serializer.hpp
@@ -6,7 +6,7 @@
 /*   By: gyong-si <gyongsi@student.42.fr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/08 10:58:01 by gyong-si          #+#    #+#             */
-/*   Updated: 2025/01/08 12:22:58 by gyong-si         ###   ########.fr       */
+/*   Updated: 2025/01/08 16:53:56 by gyong-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 #include <iostream>
 #include <string.h>
+#include <stdint.h>
 
 struct Data
 {
@@ -31,8 +32,8 @@ class Serializer
 		~Serializer(void);
 
 	public:
-		static unsigned long serialize(Data* ptr);
-		static Data* deserialize(unsigned long raw);
+		static uintptr_t serialize(Data* ptr);
+		static Data* deserialize(uintptr_t raw);
 
 };
